Protect specific paths from being deleted during cleanup
During the cleanup stage we delete blobs larger than 200 kB which do not belong
to any SL1* or SL2* tags. This removes most of the trash files, however, this
may affect a useful file modified between tagged releases. This workaround helps
to avoid removal of such files in rare cases.